Refactor endpoint request runtime helper
